refactor(activity-view): consolidate resolve_endpoint_path_entity
The path-entity preflight (UUID-parse → find-by-uuid →
raise_for_access) existed in two places:

* ``api_helpers.py::_resolve_entity`` (sc-103156) — returned a tuple
  vs ``Response`` shape for the ``/versions/`` handlers.
* ``activity.py::resolve_endpoint_path_entity`` (sc-107283) — raised
  ``PathEntityResponseError`` for the ``/activity/`` handlers.

Both flagged by clean-code-review as duplication; the activity-view's
exception-raise pattern is the cleaner of the two.

Move ``PathEntityResponseError`` + ``_RAISE_FOR_ACCESS_KWARG`` +
``resolve_endpoint_path_entity`` into ``api_helpers.py`` and rewrite
``list_versions_endpoint`` / ``get_version_endpoint`` to call it.
The ``access_kwarg`` parameter goes away — the dispatch table derives
it from ``model_cls.__name__``. Per-resource API methods become
``return list_versions_endpoint(self, Slice, uuid_str)``.

``activity.py`` re-exports the two names at the top of the module so
``charts/api.py`` / ``dashboards/api.py`` / ``datasets/api.py``
callers that already use ``activity_module.PathEntityResponseError``
don't need an import-path migration.
