+++ superset/versioning/api_helpers.py
@@ -14,22 +14,18 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Shared handlers for the ``/versions/`` REST endpoints.
+"""Shared handlers for the ``/versions/`` and ``/activity/`` REST endpoints.
 
 Each ``ChartRestApi`` / ``DashboardRestApi`` / ``DatasetRestApi`` carries
-the same three endpoint methods — ``list_versions``, ``get_version``,
-``restore_version`` — whose bodies are byte-for-byte identical apart
-from the model class, the ``security_manager.raise_for_access`` kwarg,
-and the resource-specific exception triplet on the restore path.
+the same three ``/versions/`` endpoint methods — ``list_versions``,
+``get_version``, ``restore_version`` — plus the ``activity`` endpoint
+on each resource. The bodies were byte-for-byte identical apart from
+the model class, the ``security_manager.raise_for_access`` kwarg, and
+the resource-specific exception triplet on the restore path.
+
 Extracting the bodies here lets each per-resource method collapse to
 a single delegation call, while the OpenAPI docstring + FAB decorators
 stay at the method site where they belong.
-
-The corresponding helper for the activity-view endpoint family lives
-at :func:`superset.versioning.activity.resolve_endpoint_path_entity`;
-it does only the path-entity resolution step (not the DAO + ETag
-wrapping), because the activity endpoints follow a different result
-shape.
 """
 
 from __future__ import annotations
@@ -71,49 +67,83 @@ class RestoreEndpointSpec:
     resource_label: str
 
 
-def _resolve_entity(
-    api: Any,
-    model_cls: type[Model],
-    uuid_str: str,
-    access_kwarg: str,
-) -> tuple[Any, UUID] | Response:
-    """Parse the path UUID, look up the live entity, run the read-access
-    gate.
+# Maps the versioned model class name to the keyword argument
+# ``security_manager.raise_for_access`` expects for the per-resource
+# gate. Slice → ``chart=``, Dashboard → ``dashboard=``, SqlaTable →
+# ``datasource=``. Centralised here so /versions/ and /activity/
+# endpoints share one source of truth for the dispatch.
+_RAISE_FOR_ACCESS_KWARG: dict[str, str] = {
+    "Slice": "chart",
+    "Dashboard": "dashboard",
+    "SqlaTable": "datasource",
+}
 
-    Returns ``(entity, entity_uuid)`` on success or a pre-built
-    ``Response`` (400 / 403 / 404) that the caller should return
-    directly. The split shape keeps the call site terse and lets the
-    three handler functions share the preflight without each repeating
-    the try / except dance.
+
+class PathEntityResponseError(Exception):
+    """Carries a pre-built error ``Response`` from
+    :func:`resolve_endpoint_path_entity`. Endpoints catch it and return
+    the carried response directly. The shape exists so the
+    UUID-parse + find-by-uuid + read-access check can live in one
+    place across the ``/versions/`` and ``/activity/`` endpoint
+    families."""
+
+    def __init__(self, response: Any) -> None:
+        super().__init__("PathEntityResponseError")
+        self.response = response
+
+
+def resolve_endpoint_path_entity(api: Any, model_cls: type, uuid_str: str) -> Any:
+    """Run the standard path-entity preflight for a /versions/ or
+    /activity/ endpoint:
+
+    1. Parse *uuid_str* into a UUID (or raise → 400).
+    2. Look up the live entity via ``VersionDAO.find_active_by_uuid``
+       (or raise → 404).
+    3. Run ``security_manager.raise_for_access`` with the resource-typed
+       kwarg (or raise → 403).
+
+    Returns the live entity on success. Raises
+    :class:`PathEntityResponseError` carrying the appropriate error
+    Response on any failure; the endpoint method should::
+
+        try:
+            entity = resolve_endpoint_path_entity(self, Dashboard, uuid_str)
+        except PathEntityResponseError as exc:
+            return exc.response
+
+    *api* is the FAB ``ModelRestApi`` instance — we call
+    ``api.response_400`` / ``api.response_403`` / ``api.response_404``
+    on it. Pass ``self`` from the endpoint method.
     """
     try:
         entity_uuid = UUID(uuid_str)
-    except ValueError:
-        return api.response_400(message="Invalid UUID")
+    except ValueError as exc:
+        raise PathEntityResponseError(api.response_400(message="Invalid UUID")) from exc
 
     entity = VersionDAO.find_active_by_uuid(model_cls, entity_uuid)
     if entity is None:
-        return api.response_404()
+        raise PathEntityResponseError(api.response_404())
 
+    kwarg = _RAISE_FOR_ACCESS_KWARG[model_cls.__name__]
     try:
-        security_manager.raise_for_access(**{access_kwarg: entity})
-    except SupersetSecurityException:
-        return api.response_403()
+        security_manager.raise_for_access(**{kwarg: entity})
+    except SupersetSecurityException as exc:
+        raise PathEntityResponseError(api.response_403()) from exc
 
-    return entity, entity_uuid
+    return entity
 
 
 def list_versions_endpoint(
     api: Any,
     model_cls: type[Model],
     uuid_str: str,
-    access_kwarg: str,
 ) -> Response:
     """Body of ``GET /api/v1/{resource}/<uuid>/versions/``."""
-    resolved = _resolve_entity(api, model_cls, uuid_str, access_kwarg)
-    if isinstance(resolved, Response):
-        return resolved
-    entity, entity_uuid = resolved
+    try:
+        entity = resolve_endpoint_path_entity(api, model_cls, uuid_str)
+    except PathEntityResponseError as exc:
+        return exc.response
+    entity_uuid = UUID(uuid_str)
 
     versions = VersionDAO.list_versions(model_cls, entity_uuid, entity=entity)
     if versions is None:
@@ -131,13 +161,13 @@ def get_version_endpoint(
     model_cls: type[Model],
     uuid_str: str,
     version_uuid_str: str,
-    access_kwarg: str,
 ) -> Response:
     """Body of ``GET /api/v1/{resource}/<uuid>/versions/<version_uuid>/``."""
-    resolved = _resolve_entity(api, model_cls, uuid_str, access_kwarg)
-    if isinstance(resolved, Response):
-        return resolved
-    entity, entity_uuid = resolved
+    try:
+        entity = resolve_endpoint_path_entity(api, model_cls, uuid_str)
+    except PathEntityResponseError as exc:
+        return exc.response
+    entity_uuid = UUID(uuid_str)
 
     try:
         version_uuid = UUID(version_uuid_str)
@@ -166,10 +196,11 @@ def restore_version_endpoint(
 ) -> Response:
     """Body of ``POST /api/v1/{resource}/<uuid>/versions/<version_uuid>/restore``.
 
-    Does not use :func:`_resolve_entity` — the restore command runs
-    its own ownership / existence checks via ``raise_for_ownership``
-    in ``BaseRestoreVersionCommand.validate`` and turns failures into
-    the resource-specific exception triplet packed in *spec*.
+    Does not use :func:`resolve_endpoint_path_entity` — the restore
+    command runs its own ownership / existence checks via
+    ``raise_for_ownership`` in ``BaseRestoreVersionCommand.validate``
+    and turns failures into the resource-specific exception triplet
+    packed in *spec*.
     """
     try:
         entity_uuid = UUID(uuid_str)
